+++ src/client/hooks/useThemaMenuItems.test.ts
@@ -1,8 +1,5 @@
 import { ThemaMenuItem } from '../../universal/config';
-import {
-  getThemaMenuItemsAppState,
-  isThemaActive,
-} from '../../universal/helpers/themas';
+import { getThemaMenuItemsAppState, isThemaActive } from '../config/themas';
 
 describe('useThemaMenuItems', () => {
   test('isThemaActive (No AppState Value)', () => {
